Fixing Email Verification Token Retrieval Issue
Changing the value msg_from (where the verification email is sent from) is compared to. Instead of "info@twitter.com", it should check for "info@x.com" as the verification code seems to be sent from this email now.

Fixes the "Error in LoginAcid: Email code timeout" error in my case.