Fixing and improving
Changed the way of checking for email code, its not based on <time right now, its based on latest email and checking every email for code then trying it (because if x sent a code once it wont send it again while code being still valid) <- in future there should be code exceptions if there are 2 codes to try one if wrong wait for another, im going to look into it in the future.

Preventing scraper to close on unknown errors that can occur frequently. Added 2 new error handlings.